fontconfig libX11 libXmu libxcb mesa xserver git update 10 Sep 2013
xserver          commit 47ff382d1fce25a8b097d45b79489e891f1f1228
libxcb           commit f1405d9fe4a6ddcae24585ba254389a4c4f4c8c9
libX11           commit cb107760df33ffc8630677e66e2e50aa37950a5c
libXmu           commit 2539e539eafdac88177c8ee30b043c5d52f017e4
fontconfig       commit a61e145304da86c8c35b137493bbd8fd5dd1e7f5
mesa             commit 395b9410860371a64d6b5f2d50679f29eb41729e

+++ xorg-server/hw/xwin/wintrayicon.c
@@ -170,7 +170,7 @@ winHandleIconMessage(HWND hwnd, UINT message,
             RemoveMenu(hmenuTray, ID_APP_HIDE_ROOT, MF_BYCOMMAND);
         }
 
-        SetupRootMenu((unsigned long) hmenuTray);
+        SetupRootMenu(hmenuTray);
 
         /*
          * NOTE: This three-step procedure is required for
